refactor: moved terminal.go from aeternity/ to cmd/, getErrorReason() was only used by api.go so moved there instead
