+++ cmd/terminal_test.go
@@ -1,4 +1,4 @@
-package aeternity
+package cmd
 
 import (
 	"fmt"
